ci: Split up installer tests
